fix-1: 點擊輪播下方按鈕切換照片問題
	問題：景點照片超過9張，第十張之後照片未按照預期方式切換
	原因： 只取到 class 中的第一位數字，因此第十張之後數字皆爲 1
	解決方式：改使用 split 切割字串取陣列，取得完整數字

refactor: api 改使用 blueprint 方式構建，並獨立出 models 檔案
	原本寫法：
	將 /api 下的 route都寫在 app.py 中，
	並將所有判斷式跟資料庫取值都寫在一起

	可能造成的問題：
	1. app.py 檔案太冗長，無法查看完整路由
	2. 資料庫相關操作與其他程式碼混雜，不好維護

	調整項目：
	1. 使用 blueprint 獨立出 /api 相關 route
	2. 將資料庫操作都集中在 model 中
