+++ .gitignore
@@ -142,3 +142,4 @@ cython_debug/
 note.py
 password.py
 python.py
+.data/
