+++ templates/components/sign_up.html
@@ -11,7 +11,7 @@
 			<button class="Btn buttonRegular" type="submit">註冊新帳戶</button>
 			<div class="goToText bodyMedium">
 				已經有帳戶了？
-				<a class="goToBtn goToSignIn bodyMedium" href="#">點此登入</a>
+				<a class="goToBtn goToSignIn bodyMedium">點此登入</a>
 			</div>
 		</form>
 	</div>
